feat(webpack): pass "target" to webpack config loader
This change will pass the "target" argument (can be one of "browser" or
"server") to the config loader, to allow to create a webpack
config for a specific mixin type. It also adds three new hooks, namely
"getWebpackBuildConfig", "getWebpackDevelopConfig" and
"getWebpackNodeConfig", to get the webpack config for the corresponding
target.