feat(webpack): support ESM via ".mjs" ext and "module" package field
This is also closer to the default that webpack ships and that is being
used by other tools, such as CRA.

https://github.com/webpack/webpack/blob/v4.23.1/lib/WebpackOptionsDefaulter.js#L327-L348

We previously removed this support, because it broke a couple of things,
most importantly apollo and graphql modules.
But the passed time and the knowledge that CRA and webpack are doing it
too, makes us confident in enabling it again.