feat(express): add graceful exception handling
This commit introduces a new config option: `gracePeriod`. In case of
uncaught asynchronous middleware errors or upon receival of a
`SIGTERM` signal, the server's `close` method will now be called before
exiting the process. If the server fails to finish answering pending
requests before `gracePeriod` is over, it will be terminated.