fix(react): use `require` instead of `import`
...when the `resolveAbsolutePaths`-option is set to `true`.

Also do not include the `moduleId`-property in that case since it is only needed
for the module-ID-related resolving scenario in the context of Webpack.

This fix enables downstream projects to use the result of the Babel plugin
with Node v8. Otherwise this would lead to the syntax error `Unexpected token import`.