feat(webpack): add node hot reloading on win32
This reverts to the previous implementation using hot app reloading
instead of the current module reloading. We will have to monitor
performance implications and deal with potential problems with specific
packages such as mobx. These will probably have to be excluded from
bundling using a custom mixin that implements configureBuild.