fix(webpack): use async-node target to circumvent Node 11.11.0 bug
Node v11.11.0 introduced a bug where re-requiring a file which
previously did not exist won't resolve to that file.
Therefore we switch to the async-node target of Webpack for our server-
side build which uses the vm and fs modules to load chunks, so that we
don't run into the above mentioned bug.

Read more: https://github.com/nodejs/node/issues/26926