fix(webpack): skip bootstrap in child process to skip doctor checks
Executing the bootstrap hook in the forked child process sets up the
doctor mixin to run the diagnose hook for all mixins and print out the
diagnose warnings multiple times (for each sub process).
By skipping the bootstrap hook in the child process, the doctor will
never be called and we won't get the same warnings multiple times.