fix(info): configure loglevel in LogMixin constructor
We have a few exposed APIs that just call
`initialize(overrides, options)` on `@untool/core` and in order to be
able to set the loglevel correctly, we need to handle it in the
constructor of the LogMixin too, because our APIs don't call the
`handleArguments` mixin hook.

By calling `handleArguments` in the constructor we execute just the one
`handleArguments` method that is defined in that class and are not
executing the mixin hook which would call all `handleArguments` methods
on all mixins.