refactor: consolidate "build" / "develop" webpack configs into "browser"
The separation of `build`, `develop` and `node` mixes different concerns
such as environment and targets.
Therefore we want to consolidate `build` and `develop` into a single
`browser` config, which looks at `NODE_ENV` to determine whether it is a
development build or production build.

BREAKING CHANGE: This changes the signature of the `configureBuild` hook
so that now the `target` argument can either be `browser` or `node`.
Therefore the distinction between a development build and a production
build which was previously made through `target = develop` or
`target = build` must now be done by reading the `NODE_ENV` environment
variable.

Co-authored-by: dmbch <daniel@dmbch.net>