refactor(webpack): rename target "node" to "server"
Since we now only have two webpack configs (`browser` and `node`), we
decided that a cosmetic name change from `node` to `server` would make
sense and improve the readability of our code and mixins.
Therefore we renamed the webpack config and also the `target` argument
of the `configureBuild` mixin hook.

BREAKING CHANGE: This renames the `target` argument of `configureBuild`
from `node` to `server`.

Co-authored-by: dmbch <daniel@dmbch.net>