refactor(webpack): remove obsolete "collectBuildConfigs" mixin hook
This hook was used primarily in internal APIs to collect all webpack
build configs.
But since we now always execute server builds in a child process, we
have only the browser build config in the main process and therefore
we have opted to remove this obsolete mixin hook.

BREAKING CHANGE: This removes the `collectBuildConfigs` hook.
If you have relied on this hook, try to use the child process compiler
middleware or get in touch with us to figure out an alternative.

Co-authored-by: dmbch <daniel@dmbch.net>