test: add loader unit tests
This tests the basic config loader part of untool.
Since it is not possible to call the loader with a root (cwd)
directory and cosmiconfig is always starting from current
cwd, we need to spawn the tests in a separate process.
This could be seen as a starting point, and after
refactoring the base this tests could be changed to not
spawn processes but run untool in process.