refactor: update to ava 3
To have the test run with ava 3 the configuration need to reflect the
new configuration.

Most of the current configuration complies with the default of ava 3 and
is therefore removed. But by default ava does not ignore files with
mixed patterns for double underscore and non-underscore (e.g. __tests__
and fixtures). Therefore we need to ignore these.
The throws assertion is updated as well.